+++ tests/WorkspaceFileResolver.test.cpp
@@ -41,4 +41,44 @@ TEST_CASE("resolveModule handles LocalPlayer StarterGear")
     CHECK_EQ(resolved->name, "game/StarterPack/GearScript");
 }
 
+TEST_CASE("resolveModule handles FindFirstChild")
+{
+    WorkspaceFileResolver fileResolver;
+
+    Luau::ModuleInfo baseContext{"game/ReplicatedStorage"};
+
+    // :FindFirstChild("Testing")
+    std::string tempString = "Testing";
+    Luau::AstArray<char> testingStr{tempString.data(), tempString.size()};
+    std::vector<Luau::AstExpr*> tempArgs{Luau::AstExprConstantString(Luau::Location(), testingStr).asExpr()};
+    Luau::AstArray<Luau::AstExpr*> args{tempArgs.data(), tempArgs.size()};
+    auto expr = Luau::AstExprCall(Luau::Location(),
+        Luau::AstExprIndexName(Luau::Location(), nullptr, Luau::AstName("FindFirstChild"), Luau::Location(), Luau::Position(0, 0), ':').asExpr(),
+        args, true, Luau::Location());
+    auto resolved = fileResolver.resolveModule(&baseContext, &expr);
+
+    REQUIRE(resolved.has_value());
+    CHECK_EQ(resolved->name, "game/ReplicatedStorage/Testing");
+}
+
+TEST_CASE("resolveModule fails on FindFirstChild with recursive enabled")
+{
+    WorkspaceFileResolver fileResolver;
+
+    Luau::ModuleInfo baseContext{"game/Players/LocalPlayer/StarterGear"};
+
+    // :FindFirstChild("Testing", true)
+    std::string tempString = "Testing";
+    Luau::AstArray<char> testingStr{tempString.data(), tempString.size()};
+    std::vector<Luau::AstExpr*> tempArgs{
+        Luau::AstExprConstantString(Luau::Location(), testingStr).asExpr(), Luau::AstExprConstantBool(Luau::Location(), true).asExpr()};
+    Luau::AstArray<Luau::AstExpr*> args{tempArgs.data(), tempArgs.size()};
+    auto expr = Luau::AstExprCall(Luau::Location(),
+        Luau::AstExprIndexName(Luau::Location(), nullptr, Luau::AstName("FindFirstChild"), Luau::Location(), Luau::Position(0, 0), ':').asExpr(),
+        args, true, Luau::Location());
+    auto resolved = fileResolver.resolveModule(&baseContext, &expr);
+
+    CHECK_FALSE(resolved.has_value());
+}
+
 TEST_SUITE_END();